+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## 0.1.1 (unreleased)
 
-- Retry on 404 responses for POST and PUT requests
+- Retry on 404 responses
 
 ## 0.1.0